OnBlur throws exceptions if the notetype is snippet -> Fixes #1962

+++ browser/main/lib/dataApi/attachmentManagement.js
@@ -232,39 +232,41 @@ function deleteAttachmentFolder (storageKey, noteKey) {
  * @param noteKey NoteKey of the current note. Is used to determine the belonging attachment folder.
  */
 function deleteAttachmentsNotPresentInNote (markdownContent, storageKey, noteKey) {
-  const targetStorage = findStorage.findStorage(storageKey)
-  const attachmentFolder = path.join(targetStorage.path, DESTINATION_FOLDER, noteKey)
-  const attachmentsInNote = getAttachmentsInContent(markdownContent)
-  const attachmentsInNoteOnlyFileNames = []
-  if (attachmentsInNote) {
-    for (let i = 0; i < attachmentsInNote.length; i++) {
-      attachmentsInNoteOnlyFileNames.push(attachmentsInNote[i].replace(new RegExp(STORAGE_FOLDER_PLACEHOLDER + escapeStringRegexp(path.sep) + noteKey + escapeStringRegexp(path.sep), 'g'), ''))
-    }
-  }
-
-  if (fs.existsSync(attachmentFolder)) {
-    fs.readdir(attachmentFolder, (err, files) => {
-      if (err) {
-        console.error("Error reading directory '" + attachmentFolder + "'. Error:")
-        console.error(err)
-        return
+  if (storageKey && noteKey && markdownContent !== null && typeof markdownContent !== 'undefined') {
+    const targetStorage = findStorage.findStorage(storageKey)
+    const attachmentFolder = path.join(targetStorage.path, DESTINATION_FOLDER, noteKey)
+    const attachmentsInNote = getAttachmentsInContent(markdownContent)
+    const attachmentsInNoteOnlyFileNames = []
+    if (attachmentsInNote) {
+      for (let i = 0; i < attachmentsInNote.length; i++) {
+        attachmentsInNoteOnlyFileNames.push(attachmentsInNote[i].replace(new RegExp(STORAGE_FOLDER_PLACEHOLDER + escapeStringRegexp(path.sep) + noteKey + escapeStringRegexp(path.sep), 'g'), ''))
       }
-      files.forEach(file => {
-        if (!attachmentsInNoteOnlyFileNames.includes(file)) {
-          const absolutePathOfFile = path.join(targetStorage.path, DESTINATION_FOLDER, noteKey, file)
-          fs.unlink(absolutePathOfFile, (err) => {
-            if (err) {
-              console.error("Could not delete '%s'", absolutePathOfFile)
-              console.error(err)
-              return
-            }
-            console.info("File '" + absolutePathOfFile + "' deleted because it was not included in the content of the note")
-          })
+    }
+
+    if (fs.existsSync(attachmentFolder)) {
+      fs.readdir(attachmentFolder, (err, files) => {
+        if (err) {
+          console.error("Error reading directory '" + attachmentFolder + "'. Error:")
+          console.error(err)
+          return
         }
+        files.forEach(file => {
+          if (!attachmentsInNoteOnlyFileNames.includes(file)) {
+            const absolutePathOfFile = path.join(targetStorage.path, DESTINATION_FOLDER, noteKey, file)
+            fs.unlink(absolutePathOfFile, (err) => {
+              if (err) {
+                console.error("Could not delete '%s'", absolutePathOfFile)
+                console.error(err)
+                return
+              }
+              console.info("File '" + absolutePathOfFile + "' deleted because it was not included in the content of the note")
+            })
+          }
+        })
       })
-    })
-  } else {
-    console.info("Attachment folder ('" + attachmentFolder + "') did not exist..")
+    } else {
+      console.info("Attachment folder ('" + attachmentFolder + "') did not exist..")
+    }
   }
 }
